build: import gnulib's syntax-check make rules
We're going to be eliminating autotools and gnulib, but we still wish to
have the 'make syntax-check' functionality.

This imports the minimal set of gnulib files required to keep this
working. The GNUmakefile and maint.mk files had a few tweaks to enable
them to work independently of automake generated Makefiles.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20191003105119.10698-2-berrange@redhat.com>