qemu_command: ignore non-assigned PCI multifunction hostdevs
Bulding up in the new attribute introduced by the previous patch,
let's not assign a hostdev to a QEMU domain by avoiding the
'-device' line to be created at all inside qemu_command.c.

Since the device isn't going to be guest visible, makes little
sense to assign alias and address to it, so let's also prevent
that.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20191007211136.16730-4-danielhb413@gmail.com>