virhostdev: all functions of a PCI multifunction dev must be added
This is the last piece for a full fledged PCI multifunction
partial assignment support. Now that we have a defined XML format
and QEMU is ignoring the existence of the hostdev that isn't
being assigned, let's enforce that all functions must be declared
in the domain XML for such devices. This will make sure that
we'll always detach/re-attach the whole IOMMU and will force
the user to be aware of what is going to be detached from
the host.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20191007211136.16730-5-danielhb413@gmail.com>