qemu: fix type of default video device
If a graphics device is added to XML that has no video device, libvirt
automatically adds a video device which is always of type 'cirrus', even if
the underlying qemu doesn't support cirrus.

This patch refines a bit the decision about the type of the video device.
Based on QEMU capabilities, cirrus is still preferred but only added if QEMU
supports it, otherwise QXL is preferred for SPICE and Bochs for UEFI guests.
VGA is used as a fallback.

(For some info about QEMU display devices see e.g.
https://www.kraxel.org/blog/2019/09/display-devices-in-qemu/ .)

https://bugzilla.redhat.com/show_bug.cgi?id=1668141

Signed-off-by: Pavel Mores <pmores@redhat.com>
Message-Id: <20191008111603.17081-2-pmores@redhat.com>