qemu: Add support for pseries 'host-model' machine parameter
Since QEMU v4.0.0, the host's model is no longer automatically exposed
to the guest via /proc/device-tree/host-model. There is now a machine
option 'host-model' that allows an arbitrary string to be used as the
host model [1].

This patch adds support for exposing the real host model string from
/proc/cpuinfo (also found on /proc/device-tree/model) to the guest via
-machine pseries,host-model=<model> by setting in the domain XML file:

<features>
    <host_model_passthrough state='on'/>
<features/>

The functionality is disabled by default to avoid leaking host
information to the guest.

1 - https://git.qemu.org/?p=qemu.git;a=commitdiff;h=0a794529bd11

Signed-off-by: Fabiano Rosas <farosas@linux.ibm.com>
Message-Id: <20191008164657.4286-3-farosas@linux.ibm.com>