src: rewrite driver name checker in Python
As part of an goal to eliminate Perl from libvirt build tools,
rewrite the check-drivername.pl tool in Python.

This was mostly a straight conversion, manually going line-by-line
to change the syntax from Perl to Python. Thus the overall structure
of the file and approach is the same.

In testing though it was discovered the existing code was broken
since it hadn't been updated after driver.h was split into many
files. Since the old code is being thrown away, the fix was done
as part of the rewrite rather than split into a separate commit.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20191009113741.5002-14-berrange@redhat.com>