qemu: hotplug: update device list on device deleted event
I guess this is the missing piece for [1]. It did not hurt before (like
we didn't even see any errors/warns in logs) because in
qemuProcessUpdateDevices function virDomainDefFindDevice does not find
device deleted from libvirt config. But now in case of unpluggind usb
device from host we leave device in config and thus needlessly try to
call qemuDomainRemoveDevice second time.

[1] 0dfb8a1b9: qemu: Unplug devices that disappeared when libvirtd was down

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <20191015142956.30963-9-nshirokovskiy@virtuozzo.com>