qemu_driver: use VIR_AUTOFREE() with strings 1/3
Using VIR_AUTOFREE() in all strings of qemu_driver.c make the code
a bit tidier and smaller, sparing VIR_FREE() calls and sometimes a
whole 'cleanup' label.

This is a huge change due to the amount of char * declared in
this file, thus let's split it in 3. This is the first part.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20191015200852.15778-7-danielhb413@gmail.com>