conf/domain_conf: use virStringParseYesNo helper
This helper performs a conversion from a "yes|no" string
to a corresponding boolean, and several conversions were
already done, but there are still some omissions.

For most of the remaining usages in domain_conf.c only
"yes" is explicitly checked for. This means all other
values are implicitly handled as 'false'. In this case,
use virStringParseYesNo, but if it returns -1, don't
raise an error but set the bool value to false.

Cc: crobinso@redhat.com
Cc: berrange@redhat.com
Cc: abologna@redhat.com
Cc: laine@laine.org
Cc: phrdina@redhat.com
Cc: mkletzan@redhat.com
Cc: g.sho1500@gmail.com

Signed-off-by: Mao Zhongyi <maozhongyi@cmss.chinamobile.com>
Signed-off-by: Zhang Shengju <zhangshengju@cmss.chinamobile.com>
Message-Id: <20191016023950.1623751-2-maozhongyi@cmss.chinamobile.com>