qemu_driver: improve comparison/baseline error reporting
Providing an erroneous CPU definition in the XML file provided to the
hypervisor-cpu-compare/baseline command will result in a verbose
internal error. Let's add some sanity checking before executing the QMP
commands to provide a cleaner message if something is wrong with the
CPU model or features.

An internal error will still appear for other messages originating from
QEMU e.g. if a newer feature is not available for an older CPU model.

Signed-off-by: Collin Walling <walling@linux.ibm.com>
Message-Id: <1571250061-18645-1-git-send-email-walling@linux.ibm.com>
Reviewed-by: Daniel Henrique Barboza <danielhb413@gmail.com>