util: set bridge device MAC address explicitly during virNetDevBridgeCreate
Remember when the MAC address of libvirt-created bridges weren't
stable, and changed as guests were started and stopped? Pepperidge
Farms remembers.

(No, I would never actually push a comment like that! Just wanted to
get your attention).

When libvirt first implemented a stable and configurable MAC address
for the bridges created for libvirt virtual networks (commit
5754dbd56d, in libvirt v8.8.0) most distro stable releases didn't
support explicitly setting the MAC address of a bridge; the bridge
just always assumed the lowest numbered MAC of all attached
interfaces. Because of this, we stabilized the bridge MAC address by
creating a "dummy" tap interface with a MAC address guaranteed to be
lower than any of the guest tap devices' MACs (which all started with
0xFE, so it's not difficult to do) and attached it to the bridge -
this was the inception of the "virbr0-nic" device that has confused so
many people over the years.

Even though the linux kernel had recently gained support for
explicitly setting a bridge MAC, we deemed it unnecessary to set the
MAC that way, because the other (indirect) method worked everywhere.

But recently there have been reports that the bridge MAC address was
not following the setting in the network config, and mismatched the
MAC of the dummy tap device (which was still correct). It turns out
that this is due to a change in systemd/udev that persists whatever
MAC address is set for a bridge when it's initially started:

  https://github.com/systemd/systemd/blob/master/NEWS#L499

which was the result of:

  https://github.com/systemd/systemd/issues/3374

(apparently if there is no MAC saved for a bridge by the name of a
bridge being created, the random MAC generated during creation is
saved, and then that same MAC is used to explicitly set the MAC each
time it is created). Once a bridge has an explicitly set MAC, the "use
the lowest numbered MAC of attached devices" rule is ignored, so our
dummy tap device is like the goggles - it does nothing! (well, almost).

We could whine about changes in default behavior, etc. etc., but
because the change was in response to actual user problems, that seems
likely a fruitless task. Fortunately, time has marched on, and even
distro releases that are old enough that they are no longer supported
by upstream libvirt (e.g. RHEL6) have support for explicitly setting a
bridge device MAC address, either during creation or with a separate
ioctl after creation, so we can now do that.

The method is to add a mac arg to virNetDevBridgeCreate(). In the case
of platforms where the bridge is created with a netlink RTM_NEWLINK
message, we just add the desired mac to the message. For platforms
that still use an ioctl (either SIOCBRADDBR or SIOCIFCREATE2), we make
a separate call to virNetDevSetMAC() after creating the bridge.

This makes the dummy tap pointless for purposes of setting the MAC
address, but it is still useful for IPv6 DAD initialization (which
apparently requires at least one interface to be attached to the
bridge and online), so it hasn't been removed. (I'm considered making
another patch to add the dummy tap device only if the network needs
IPv6 DAD, but haven't decided yet if it's worthwhile).

(NB: we can safely *always* call virNetDevBridgeCreate() with
&def->mac from the network driver because, in spite of the existence
of a "mac_specified" bool in the config suggesting that it may not
always be present, in reality a mac address will always be added to
any network that doesn't have one - this is guaranteed in all cases by
commit a47ae7c004)

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1760851
Message-Id: <20191018015000.456596-1-laine@laine.org>