qemu: fix type of default video device
If a graphics device is added to XML that has no video device, libvirt
automatically added a video device which was always of type 'cirrus' on
x86_64, even if the underlying qemu didn't support cirrus.

This patch refines a bit the decision about the type of the video device.
Based on QEMU capabilities, cirrus is still preferred but only added if
QEMU supports it, otherwise VGA is used if supported by QEMU.  There is now
no fallback as libvirt only aspires to generate a basic working config and
leaves anything more specific up to higher-level management tools.

https://bugzilla.redhat.com/show_bug.cgi?id=1668141
Signed-off-by: Pavel Mores <pmores@redhat.com>
Message-Id: <20191018102722.1934-2-pmores@redhat.com>