qemu: add tests of the default video type selection algorithm
The test case for x86_64 and neither cirrus nor vga capability is of the
xml2argv type because it actually fails to parse the XML at all [*] which
is something that xml2xml tests don't seem to handle.  xml2argv test fails
to produce a qemu argv for this case which xml2argv tests can handle.

The (unrelated) "graphics-egl-headless-rendernode" had to be adjusted as it
expects cirrus to be added as per previous behaviour.  It now has to use
QEMU_CAPS_DEVICE_CIRRUS_VGA to get that (otherwise it'd end up with vga).

[*] This is a consequence of the decision not to have a fallback if the
obvious choices (cirrus and vga) aren't viable due to missing QEMU caps.

Signed-off-by: Pavel Mores <pmores@redhat.com>
Message-Id: <20191018102722.1934-3-pmores@redhat.com>