qemu: domain: Split out setup of virStorageSource from qemu driver config
qemuDomainPrepareDiskSourceData historically prepared everything but
we'we split out the majority of the functionality so that it sets up
predominantely only according to the configuration of the disk. There
was one leftover bit of setting the gluster debug level from the config.

Split this out into a separate function so that
qemuDomainPrepareDiskSourceData only perpares based on the disk.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <9d926b1109df55ca23b345ee0314f2d8ceec6ce9.1571414890.git.pkrempa@redhat.com>