conf: Do not validate resolution XML if 'x' or/and 'y' are 0.
There is an issue with <resolution/> when of 'x' or 'y' settings are 0.
Function virDomainVideoResolutionDefParseXML() will validate this XML,
but both 'x' and 'y' will be removed. One example, if someone defines
this settings:

    <model ...>
      <resolution x='1024' y='0'/>
    <model/>

After applying this settings, funcion libvirt will remove both
resolutions because virDomainVideoResolutionDefFormat() requires 'x' and
'y' higher than 0. So, the example above will become:

    <model ...>
      <resolution/>
    <model/>

Now, libvirt only adds <resolution/> with 'x' and 'y' if boths strings
are not NULL AND they are higher than 0.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20191020034520.29721-1-jcfaracco@gmail.com>