XML definitions for guest vNUMA and parsing routines
This patch adds XML definitions to a guest with a vNUMA layout and
contains routines to parse the same.  The guest vNUMA specification
looks like:

  <vnuma mode='host|node'
         distribution='contiguous|siblings|round-robin|interleave'>
    <memory unit='#unitsize'>size</memory>
    <partition nodeset='#nodes' cells='#cells'/>
  </vnuma>

With mode='host' the guest XML is rendered to match the host's NUMA
topology.

With mode='node' the guest XML is rendered according to the "nodes"
and "cells" attributes of the <partition> element.

Signed-off-by: Wim ten Have <wim.ten.have@oracle.com>
Message-Id: <20191021192108.25974-2-wim.ten.have@oracle.com>