Add --force flag to virsh blockresize
This commit simply adds the '--force' flag to the blockresize
command to prevent accidental shrinking of a block device. Similar
behaviour is already present on the vol-resize command and it
makes sense to mimic it.

Implemented in virsh and not in the API. This is to prevent existing
applications using the BlockResize-API from 'breaking'.

Signed-off-by: Patrik Martinsson <martinsson.patrik@gmail.com>
Message-Id: <20191021195539.1623177-2-martinsson.patrik@gmail.com>