remote: dispatch: Remove return value from make_nonnull_* helpers
After convertion to g_strdup, the helpers now always return success.
Remove the return value to simplify the callers.

Note that many occurences of these is in the code generated by
gendispatch.pl. Since gendispatch aggregates many cases together a
incremental conversion would require more invasive changes to
gendispatch for the time of coversion which doesn't make sense.

Also in many cases the helper was the last place where the 'error:'
label was used and thus also those conversions must be included in this
patch.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <a0d140d325229ad034c4bdc9d4f27a13f6b9c5ef.1571754966.git.pkrempa@redhat.com>