conf: report errors when parsing video resolution
The current code doesn't properly handle errors when parsing a video
device's resolution.

This patch changes the parse function signature to return an error
when we're missing an 'x' or 'y' parameter or when the 'x' or 'y'
parameters are not positive integers. No error is returned when no
'resolution' element is found.

Previously we were returning a NULL structure for the case where 'x' or
'y' were missing, but were returning an under-specified structure for
the other error cases.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20191023174649.28887-5-jjongsma@redhat.com>