conf: report errors when parsing video acceleration
In order to differentiate between a configuration that does not specify
any video acceleration and one that is incorrectly specified, change the
signature of virDomainVideoAccelDefParseXML() to return an error
response.

If any of the values are invalid, report an error rather than returning
a partially-specified accel object. If no 'acceleration' node is found,
do not report an error since it is optional.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20191023174649.28887-6-jjongsma@redhat.com>